Actually, this code fits better in util.
